+++ lib/proxy/proxy.dart
@@ -1,4 +1,5 @@
 export 'axis.dart';
 export 'button.dart';
+export 'listTile.dart';
 export 'radio.dart';
 export 'textFormField.dart';